Move Relu and SoftMax Op declarations into header files so they can be registered by experimental devices.
Right now tensorflow/core/kernels explicitly depends on all Eigen devices that might want to implement any of the templated Eigen Ops. This is because the template classes that need to be specialized are defined in .cc files, so the specializations themselves have to appear there too. Moving the classes to .h files allows us to use arbitrary Eigen devices defined outside of tensorflow/core, which fits better with the intent behind core/kernels.

Over time more kernels may need to be refactored this way for the same reason.
Change: 117452814
