+++ tensorflow/core/kernels/BUILD
@@ -1062,6 +1062,7 @@ filegroup(
         "slice_op.h",
         "softmax_op.cc",
         "softmax_op.h",
+        "softmax_op_functor.h",
         "split_lib.h",
         "split_lib_cpu.cc",
         "split_op.cc",
@@ -1099,6 +1100,7 @@ filegroup(
         "reduction_ops.h",
         "reduction_ops_common.h",
         "relu_op.h",
+        "relu_op_functor.h",
         "save_restore_tensor.h",
         "softplus_op.h",
         "softsign_op.h",
